requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Qdrant
qdrant-client==1.7.3

# AI/ML Libraries
openai==1.10.0
transformers==4.37.0
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# Audio Processing
openai-whisper==20231117
soundfile==0.12.1
librosa==0.10.1

# Data Processing
pandas==2.2.0
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
sentence-transformers==2.3.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
